+++ xll24/src/fpx.c
@@ -2,6 +2,7 @@
 // Copyright (c) KALX, LLC. All rights reserved. No warranty made.
 #include <stdlib.h>
 #include <string.h>
+#include <stddef.h>
 #include "fpx.h"
 
 int fpx_index(struct fpx* p, int i, int j)
@@ -48,7 +49,7 @@ struct fpx* fpx_transpose(struct fpx* fpx)
 	if (r > 1 && c > 1) {
 		struct fpx* fpx_ = fpx_malloc(c, r);
 		if (fpx_) {
-			memcpy(fpx_->array, fpx->array, n * sizeof(double));
+			memcpy(fpx_->array, fpx->array, (size_t)n * sizeof(double));
 			for (int k = 1; k < n - 1; ++k) {
 				// if 0 < k < n - 1 and k = c j + i then
 				// r k % (n - 1) = (r c j + r i) % (n - 1) = j + (n - 1) j + r i % (n - 1) = j + r i
